inventory_service/pyproject.toml
[project]
name = "inventory-service"
version = "0.1.0"
description = "Inventory service for Hotel System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]",
    "sqlalchemy",
    "asyncpg",
    "psycopg2-binary",
    "pydantic-settings",
    "python-dotenv",
    "faker",
    "sentry-sdk[fastapi]>=2.31.0",
    "uvicorn[standard]",
    "python-json-logger>=3.3.0",
    "opentelemetry-sdk>=1.34.1",
    "opentelemetry-instrumentation-logging>=0.55b1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-api>=1.34.1",
]

[tool.setuptools]
package-dir = {"" = "app"}
packages = ["api", "config", "db", "models", "service"] 
